core: prepare for Starknet Gateway deprecation
**Summary**

This big refactor includes changes that are needed to move starknet-react
forward in a world without the Starknet Gateway 💀.

Notable changes:

 - introduce chains, which contain information about a specific chain.
 - introduce providers, which are a convenient way to create starknet
   `RpcProvider`.
 - connectors' data like name and icons come from configuration and
 is not loaded at runtime like before.

Since we have to make a breaking release, this commit also cleans up the API to
be more coherent. In practice, this means all react-query values returned by
hooks are now the same for all hooks in the library.